Ultima Actualización

+++ WcfService/Dominio/Pago.cs
@@ -23,7 +23,7 @@ namespace WcfService.Dominio
          public string ds_estado { get; set; }
 
          [DataMember]
-        public double qt_monto { get; set; }
+        public decimal qt_monto { get; set; }
          [DataMember]
         public DateTime dt_fecha { get; set; }
     }
